+++ mithril-aggregator/src/http_server/routes/certificate_routes.rs
@@ -86,7 +86,7 @@ mod handlers {
             Ok(None) => Ok(reply::empty(StatusCode::NO_CONTENT)),
             Err(err) => {
                 warn!("certificate_pending::error"; "error" => ?err);
-                Ok(reply::internal_server_error(err))
+                Ok(reply::server_error(err))
             }
         }
     }
@@ -104,7 +104,7 @@ mod handlers {
             Ok(certificates) => Ok(reply::json(&certificates, StatusCode::OK)),
             Err(err) => {
                 warn!("certificate_certificates::error"; "error" => ?err);
-                Ok(reply::internal_server_error(err))
+                Ok(reply::server_error(err))
             }
         }
     }
@@ -127,7 +127,7 @@ mod handlers {
             Ok(None) => Ok(reply::empty(StatusCode::NOT_FOUND)),
             Err(err) => {
                 warn!("certificate_certificate_hash::error"; "error" => ?err);
-                Ok(reply::internal_server_error(err))
+                Ok(reply::server_error(err))
             }
         }
     }